safari-600.7-branch/LayoutTests/js/order-of-operations-expected.txt

PASS (function(){ leftRight = ""; left() > right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() >= right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() < right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() <= right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() + right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() - right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() / right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() * right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() % right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() << right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() >> right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() >>> right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() || right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() && right(); return leftRight; })() is "Left"
PASS (function(){ leftRight = ""; left() & right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() | right(); return leftRight; })() is "LeftRight"
PASS (function(){ leftRight = ""; left() ^ right(); return leftRight; })() is "LeftRight"
PASS testEvaluationOfArguments() is true
PASS successfullyParsed is true

TEST COMPLETE
